Fixes the problem
My first suspicion turned out to be the right one. The menu was being paused when it became visible. Setting the menu to process allowed everything to work.

I also cleared up a few warnings you were getting.

This is a great start! Good work. I look forward to playing the final version.